refactor: switch pages to database util
