+++ app/sitemap.ts
@@ -1,14 +1,10 @@
-import { supabase } from "@/utils/supabase";
-import { PostgrestSingleResponse } from "@supabase/supabase-js";
 import { MetadataRoute } from "next";
 import { SiteConfig } from "@/lib/store";
-import { mapDataToConfig } from "@/lib/config";
-import { ConfigProp } from "@/types/config";
 import { Design } from "@/types/design";
+import { getSiteConfig, fetchDesigns } from "@/utils/database";
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
-  const { data: params } = await supabase.from("studio").select("*");
-  const newConfig: SiteConfig = mapDataToConfig(params as ConfigProp[]);
+  const newConfig: SiteConfig = await getSiteConfig();
   const lastModified: string = new Date().toISOString().split("T")[0];
   const sitemap = [
     {
@@ -43,15 +39,13 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
     },
   ];
   // Google's limit is 50,000 URLs per sitemap
-  const { data }: PostgrestSingleResponse<Design[]> = await supabase
-    .from("designs")
-    .select("*");
+  const { designs } = await fetchDesigns(1, "", "", 1000);
 
-  const designs = (data || []).map((design: Design) => ({
+  const designEntries = (designs || []).map((design: Design) => ({
     url: `https://${newConfig.domain}/designs/${design.id}`,
     lastModified: (design.createdAt || "").toString().split("T")[0],
     priority: 0.5,
     changeFreq: "monthly",
   }));
-  return [...sitemap, ...designs];
+  return [...sitemap, ...designEntries];
 }
